requirements.txt
# Core Dependencies
pandas==2.2.0
numpy==1.26.3
python-dotenv==1.0.1
pyyaml==6.0.1

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.25

# Machine Learning
lightgbm==4.3.0
scikit-learn==1.4.0
mlflow==2.10.2

# AI / Gemini
google-generativeai==0.3.2

# Dashboard
streamlit==1.31.0
plotly==5.18.0

# Utilities
tqdm==4.66.1
